fix incorrect tout=0 handling
in order to avoid divisions by zero we artificially replace 0 in tout with 1. this should not matter, since later we multiply it with the rates, which are 0 anyways. but it lead to problems with the cumulative multiplication in term2, since now we hade the wrong holding probbabilities (too low). hence the resulting galerkin projection had too small row sums.
this is fixed by computing the holding probabilities before the 0 fix